pyproject.toml
[tool.poetry]
name          = "project derailleur"
version       = "0.1.0"
description   = "Shifts gears through various data science techniques to analyze and visualize bike sharing trends, using the UCI Bike Sharing Dataset."
authors       = ["James Parkington <pongs.acreage.0o@icloud.com>"]
readme        = "README.md"

[tool.poetry.dependencies]
python        = ">=3.11,<3.13"
pandas        = "2.0.3"
numpy         = "1.25.1"
matplotlib    = "3.7.2"
seaborn       = "0.12.2"
adjusttext    = "0.8"
scipy         = "^1.11.1"
nbformat      = "^5.9.2"
pillow        = ">=9.4.0,<10.0.0"

[tool.poetry.group.dev.dependencies]
ipykernel     = "6.25.0"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"